Source Code
This is the first ever addition to the hydrogen language. At the moment, it only does addition and subtraction and returns the exit code.

You must use the format: return num + num; (space between everything) else it'll throw an error. I always code like that so thought it would be cool to make it mandatory for the language. It's riddled with bugs and only works for macos (intel chips) for now. The goal would be to add: multiplication and division next along with priority of operators.